make chunked_indices mod-private
