+++ encodings/fastlanes/src/bitpacking/compute/filter.rs
@@ -7,6 +7,7 @@ use vortex_array::{ArrayData, IntoArrayData, IntoArrayVariant};
 use vortex_dtype::{match_each_unsigned_integer_ptype, NativePType};
 use vortex_error::VortexResult;
 
+use super::chunked_indices;
 use crate::bitpacking::compute::take::UNPACK_CHUNK_THRESHOLD;
 use crate::{BitPackedArray, BitPackedEncoding};
 
@@ -102,36 +103,6 @@ fn filter_indices<T: NativePType + BitPacking + ArrowNativeType>(
     });
 
     values
-}
-
-pub fn chunked_indices<F: FnMut(usize, &[usize])>(
-    mut indices: impl Iterator<Item = usize>,
-    offset: usize,
-    mut chunk_fn: F,
-) {
-    let mut indices_within_chunk: Vec<usize> = Vec::with_capacity(1024);
-
-    let Some(first_idx) = indices.next() else {
-        return;
-    };
-
-    let mut current_chunk_idx = (first_idx + offset) / 1024;
-    indices_within_chunk.push((first_idx + offset) % 1024);
-    for idx in indices {
-        let new_chunk_idx = (idx + offset) / 1024;
-
-        if new_chunk_idx != current_chunk_idx {
-            chunk_fn(current_chunk_idx, &indices_within_chunk);
-            indices_within_chunk.clear();
-        }
-
-        current_chunk_idx = new_chunk_idx;
-        indices_within_chunk.push((idx + offset) % 1024);
-    }
-
-    if !indices_within_chunk.is_empty() {
-        chunk_fn(current_chunk_idx, &indices_within_chunk);
-    }
 }
 
 fn filter_slices<T: NativePType + BitPacking + ArrowNativeType>(
